refactor: rename -> reroute, route
