+++ src/mediaQueryContext.ts
@@ -2,10 +2,10 @@ const mobileMediaQueryList = window.matchMedia('(max-width: 559px)');
 
 let isMobile = mobileMediaQueryList.matches; // init
 
-export function mediaQueryContextProvider(reRoute: () => void): void {
+export function mediaQueryContextProvider(reroute: () => void): void {
   mobileMediaQueryList.addListener((mobileMediaQueryList) => {
     isMobile = mobileMediaQueryList.matches;
-    reRoute();
+    reroute();
   });
 }
